advisory_modules/requirement.txt
pandas
numpy
openpyxl
python-dotenv
google-generativeai
